Fix bug that was stopping mouse over/out of one mark working.
A user was having trouble with panning/resizing due to a hack I'd
added to avoid infinite mouse events triggering constant redraws.
In the end this code isn't needed as the constant redraws can be
avoided by a slightly cleverer approach to handling the mouse over
and out events in the graphs that were having the problem.